Add BCE UI framework compat for cTab Map UI
