+++ cTab/cTab_classes.hpp
@@ -1,3 +1,13 @@
+/*
+  #NOTE - This will be included in "RscTitles" + "Dialog"
+*/
+
+//- TAD
+  class cTab_TAD_OSD_navModeOrScale;
+  class cTab_TAD_OSD_txtToggleIconBg;
+  class cTab_TAD_OSD_txtToggleIcon;
+  class cTab_TAD_RscMapControl_BLACK;
+
 //-Clean up
 #if MAP_MODE > 0
 	//-Change TOPO -> Enhanced GPS
@@ -9,3 +19,7 @@
 	class cTab_microDAGR_RscMapControl: cTab_RscMapControl{};
 	class cTab_TAD_RscMapControl: cTab_RscMapControl{};
 #endif
+
+//- Phone
+  #include "cTab_Android_Widgets.hpp"
+  #include "cTab_Android_Layout.hpp"
